### :nut_and_bolt: Description
We are enabling two custom ways of sorting the service-groups:

1) **Sort by Group Health OK percentage number:** from smallest to largest.
(0-100) Within the same percentage number sort by level of criticality.
The level of criticality must always be Critical, Unknown, Warning, OK
in that order.

2) **Sort by Level of Criticality:** The order is Critical, Unknown, Warning
and Ok. Within the same level of criticality sort by “group health Ok
percentage number”. The percentage number must always be ordered
from 0-100

We are also changing the default sort field to be `percent_ok`.
Finally, a quick fix of a flaky test that was failing randomly.

### :+1: Definition of Done
1) By default, sort by Group Health OK percentage number: from smallest to largest. (0-100) Within the same percentage number sort by level of criticality. When the sorting order is toggled, within the same percentage number, still sort by level of criticality. The level of criticality must always be Critical, Unknown and Warning in that order.

2) Switch the primary sorting criteria with the secondary. That is by default, sort by Level of criticality: The order is Critical, Unknown, Warning and Ok. Within the same level of criticality sort by “group health Ok percentage number”. The group health OK percentage number must always be ordered from 0-100.

### :athletic_shoe: Demo Script / Repro Steps
Enter the studio then start, build and populate applications-service.
Verify the new default service-group order.
```
$ hab studio enter
[1][default:/src:0]# start_applications_service
[2][default:/src:0]# applications_populate_database
[3][default:/src:0]# applications_get_service_groups | grep healthPercentage
      "healthPercentage": 20,
      "healthPercentage": 20,
      "healthPercentage": 60,
      "healthPercentage": 80,
      "healthPercentage": 100,
      "healthPercentage": 100,
```
You can just play a bit more with the GRPC functions. 😄

### :chart_with_upwards_trend: Code Benchmark (MASTER Vs PR)
We are introducing some benchmarks for the applications
service at this PR https://github.com/chef/automate/pull/495 and made it official with https://github.com/chef/automate/pull/510

Here is a comparison between what is running currently in
**master versus this PR.**
```
[18][default:/src:0]# benchstat /src/benchmarks/applications-service/txt/PR_510.txt /src/benchmarks/applications-service/txt/PR_490.txt
name                                                   old time/op    new time/op    delta
IngestionNewServices-4                                   9.05ms ±14%    8.60ms ±11%     ~     (p=0.690 n=5+5)
IngestionUpdateServices-4                                3.00ms ±11%    2.59ms ± 4%  -13.61%  (p=0.008 n=5+5)
GetServiceGroups/max_1000_with_0_groups-4                2.42ms ± 2%    2.25ms ± 5%   -7.21%  (p=0.008 n=5+5)
GetServiceGroups/max_1000_with_100_groups-4              7.95ms ± 2%    6.97ms ± 2%  -12.29%  (p=0.008 n=5+5)
GetServiceGroups/max_1000_with_200_groups-4              18.0ms ± 2%    16.0ms ± 1%  -11.13%  (p=0.008 n=5+5)
GetServiceGroups/max_1000_with_300_groups-4              32.6ms ± 3%    28.5ms ± 0%  -12.53%  (p=0.008 n=5+5)
GetServiceGroups/max_1000_with_400_groups-4              57.5ms ±19%    43.7ms ± 2%  -23.97%  (p=0.008 n=5+5)
GetServiceGroups/max_1000_with_500_groups-4              75.4ms ± 3%    66.1ms ± 1%  -12.32%  (p=0.008 n=5+5)
GetServiceGroups/max_1000_with_600_groups-4               103ms ± 2%      91ms ± 1%  -10.81%  (p=0.008 n=5+5)
GetServiceGroups/max_1000_with_700_groups-4               127ms ± 6%     120ms ± 1%   -5.53%  (p=0.008 n=5+5)
GetServiceGroups/max_1000_with_800_groups-4               168ms ± 2%     152ms ± 0%   -9.40%  (p=0.008 n=5+5)
GetServiceGroups/max_1000_with_900_groups-4               205ms ± 3%     189ms ± 1%   -8.05%  (p=0.008 n=5+5)
GetServiceGroups/max_1000_with_1000_groups-4              253ms ± 3%     230ms ± 1%   -9.05%  (p=0.008 n=5+5)
GetServiceGroupsHeavyLoad/max_5000_with_0_groups-4       2.53ms ±20%    2.21ms ± 3%  -12.69%  (p=0.008 n=5+5)
GetServiceGroupsHeavyLoad/max_5000_with_1000_groups-4     255ms ± 7%     233ms ± 1%   -8.74%  (p=0.008 n=5+5)
GetServiceGroupsHeavyLoad/max_5000_with_2000_groups-4     1.15s ±16%     0.87s ± 1%  -24.21%  (p=0.008 n=5+5)
GetServiceGroupsHeavyLoad/max_5000_with_3000_groups-4     2.07s ± 1%     1.90s ± 2%   -8.26%  (p=0.008 n=5+5)
GetServiceGroupsHeavyLoad/max_5000_with_4000_groups-4     3.93s ±20%     3.38s ± 1%  -13.87%  (p=0.008 n=5+5)
GetServiceGroupsHeavyLoad/max_5000_with_5000_groups-4     5.38s ± 1%     5.26s ± 1%   -2.21%  (p=0.008 n=5+5)
GetServices/max_1000_with_0_services-4                   2.13ms ± 8%    1.97ms ± 3%   -7.93%  (p=0.016 n=5+5)
GetServices/max_1000_with_100_services-4                 2.52ms ± 5%    2.61ms ± 9%     ~     (p=0.310 n=5+5)
GetServices/max_1000_with_200_services-4                 2.80ms ± 2%    2.64ms ± 3%   -5.92%  (p=0.008 n=5+5)
GetServices/max_1000_with_300_services-4                 3.00ms ± 5%    2.90ms ± 4%     ~     (p=0.222 n=5+5)
GetServices/max_1000_with_400_services-4                 3.09ms ± 2%    3.17ms ±20%     ~     (p=0.310 n=5+5)
GetServices/max_1000_with_500_services-4                 3.36ms ± 1%    3.71ms ±14%     ~     (p=0.690 n=5+5)
GetServices/max_1000_with_600_services-4                 3.61ms ± 6%    3.88ms ± 7%   +7.56%  (p=0.032 n=5+5)
GetServices/max_1000_with_700_services-4                 3.80ms ± 2%    3.87ms ± 1%   +2.05%  (p=0.008 n=5+5)
GetServices/max_1000_with_800_services-4                 4.04ms ± 2%    4.12ms ± 2%     ~     (p=0.095 n=5+5)
GetServices/max_1000_with_900_services-4                 4.32ms ± 3%    4.49ms ± 3%   +3.89%  (p=0.016 n=5+5)
GetServices/max_1000_with_1000_services-4                4.48ms ± 3%    4.59ms ± 2%   +2.34%  (p=0.032 n=5+5)
GetServicesHeavyLoad/max_5000_with_0_services-4          1.71ms ± 2%    1.69ms ± 3%     ~     (p=0.548 n=5+5)
GetServicesHeavyLoad/max_5000_with_1000_services-4       4.60ms ± 4%    4.56ms ± 4%     ~     (p=0.690 n=5+5)
GetServicesHeavyLoad/max_5000_with_2000_services-4       7.23ms ± 5%    7.27ms ± 5%     ~     (p=0.841 n=5+5)
GetServicesHeavyLoad/max_5000_with_3000_services-4       9.63ms ± 4%    9.85ms ± 6%     ~     (p=0.548 n=5+5)
GetServicesHeavyLoad/max_5000_with_4000_services-4       12.3ms ± 2%    12.3ms ± 3%     ~     (p=0.841 n=5+5)
GetServicesHeavyLoad/max_5000_with_5000_services-4       14.7ms ± 1%    14.8ms ± 2%     ~     (p=1.000 n=5+5)
```

### There are no significant changes between them! 💯 

### :chains: Related Resources
https://chefio.atlassian.net/browse/A2-840

### :white_check_mark: Checklist

- [X] Necessary tests added/updated?
- [X] Necessary docs added/updated?
- [X] Code actually executed?
- [X] Vetting performed (unit tests, lint, etc.)?