@afiune 2019-06-07
I'm about to add the benchmarks to this PR. Just like we defined at https://github.com/chef/automate/pull/510