@afiune 2019-06-07
Computed output:
```
[17][default:/src:0]# benchstat /src/benchmarks/applications-service/txt/PR_490.txt
name                                                   time/op
IngestionNewServices-4                                 8.60ms ±11%
IngestionUpdateServices-4                              2.59ms ± 4%
GetServiceGroups/max_1000_with_0_groups-4              2.25ms ± 5%
GetServiceGroups/max_1000_with_100_groups-4            6.97ms ± 2%
GetServiceGroups/max_1000_with_200_groups-4            16.0ms ± 1%
GetServiceGroups/max_1000_with_300_groups-4            28.5ms ± 0%
GetServiceGroups/max_1000_with_400_groups-4            43.7ms ± 2%
GetServiceGroups/max_1000_with_500_groups-4            66.1ms ± 1%
GetServiceGroups/max_1000_with_600_groups-4            91.5ms ± 1%
GetServiceGroups/max_1000_with_700_groups-4             120ms ± 1%
GetServiceGroups/max_1000_with_800_groups-4             152ms ± 0%
GetServiceGroups/max_1000_with_900_groups-4             189ms ± 1%
GetServiceGroups/max_1000_with_1000_groups-4            230ms ± 1%
GetServiceGroupsHeavyLoad/max_5000_with_0_groups-4     2.21ms ± 3%
GetServiceGroupsHeavyLoad/max_5000_with_1000_groups-4   233ms ± 1%
GetServiceGroupsHeavyLoad/max_5000_with_2000_groups-4   874ms ± 1%
GetServiceGroupsHeavyLoad/max_5000_with_3000_groups-4   1.90s ± 2%
GetServiceGroupsHeavyLoad/max_5000_with_4000_groups-4   3.38s ± 1%
GetServiceGroupsHeavyLoad/max_5000_with_5000_groups-4   5.26s ± 1%
GetServices/max_1000_with_0_services-4                 1.97ms ± 3%
GetServices/max_1000_with_100_services-4               2.61ms ± 9%
GetServices/max_1000_with_200_services-4               2.64ms ± 3%
GetServices/max_1000_with_300_services-4               2.90ms ± 4%
GetServices/max_1000_with_400_services-4               3.17ms ±20%
GetServices/max_1000_with_500_services-4               3.71ms ±14%
GetServices/max_1000_with_600_services-4               3.88ms ± 7%
GetServices/max_1000_with_700_services-4               3.87ms ± 1%
GetServices/max_1000_with_800_services-4               4.12ms ± 2%
GetServices/max_1000_with_900_services-4               4.49ms ± 3%
GetServices/max_1000_with_1000_services-4              4.59ms ± 2%
GetServicesHeavyLoad/max_5000_with_0_services-4        1.69ms ± 3%
GetServicesHeavyLoad/max_5000_with_1000_services-4     4.56ms ± 4%
GetServicesHeavyLoad/max_5000_with_2000_services-4     7.27ms ± 5%
GetServicesHeavyLoad/max_5000_with_3000_services-4     9.85ms ± 6%
GetServicesHeavyLoad/max_5000_with_4000_services-4     12.3ms ± 3%
GetServicesHeavyLoad/max_5000_with_5000_services-4     14.8ms ± 2%

name                                                   alloc/op
IngestionNewServices-4                                 15.9kB ± 0%
IngestionUpdateServices-4                              9.73kB ± 0%
GetServiceGroups/max_1000_with_0_groups-4              11.4kB ± 0%
GetServiceGroups/max_1000_with_100_groups-4            37.9kB ± 0%
GetServiceGroups/max_1000_with_200_groups-4            38.1kB ± 0%
GetServiceGroups/max_1000_with_300_groups-4            38.2kB ± 0%
GetServiceGroups/max_1000_with_400_groups-4            38.2kB ± 0%
GetServiceGroups/max_1000_with_500_groups-4            38.4kB ± 0%
GetServiceGroups/max_1000_with_600_groups-4            38.3kB ± 0%
GetServiceGroups/max_1000_with_700_groups-4            38.5kB ± 0%
GetServiceGroups/max_1000_with_800_groups-4            38.4kB ± 0%
GetServiceGroups/max_1000_with_900_groups-4            38.4kB ± 0%
GetServiceGroups/max_1000_with_1000_groups-4           38.6kB ± 0%
GetServiceGroupsHeavyLoad/max_5000_with_0_groups-4     11.4kB ± 0%
GetServiceGroupsHeavyLoad/max_5000_with_1000_groups-4  38.6kB ± 0%
GetServiceGroupsHeavyLoad/max_5000_with_2000_groups-4  40.0kB ± 0%
GetServiceGroupsHeavyLoad/max_5000_with_3000_groups-4  40.4kB ± 0%
GetServiceGroupsHeavyLoad/max_5000_with_4000_groups-4  40.4kB ± 0%
GetServiceGroupsHeavyLoad/max_5000_with_5000_groups-4  40.3kB ± 0%
GetServices/max_1000_with_0_services-4                 12.3kB ± 0%
GetServices/max_1000_with_100_services-4               38.6kB ± 0%
GetServices/max_1000_with_200_services-4               38.6kB ± 0%
GetServices/max_1000_with_300_services-4               38.6kB ± 0%
GetServices/max_1000_with_400_services-4               38.6kB ± 0%
GetServices/max_1000_with_500_services-4               38.6kB ± 0%
GetServices/max_1000_with_600_services-4               38.6kB ± 0%
GetServices/max_1000_with_700_services-4               38.6kB ± 0%
GetServices/max_1000_with_800_services-4               38.6kB ± 0%
GetServices/max_1000_with_900_services-4               38.6kB ± 0%
GetServices/max_1000_with_1000_services-4              38.6kB ± 0%
GetServicesHeavyLoad/max_5000_with_0_services-4        12.3kB ± 0%
GetServicesHeavyLoad/max_5000_with_1000_services-4     38.6kB ± 0%
GetServicesHeavyLoad/max_5000_with_2000_services-4     38.6kB ± 0%
GetServicesHeavyLoad/max_5000_with_3000_services-4     38.6kB ± 0%
GetServicesHeavyLoad/max_5000_with_4000_services-4     38.6kB ± 0%
GetServicesHeavyLoad/max_5000_with_5000_services-4     38.7kB ± 0%

name                                                   allocs/op
IngestionNewServices-4                                    549 ± 0%
IngestionUpdateServices-4                                 403 ± 0%
GetServiceGroups/max_1000_with_0_groups-4                 427 ± 0%
GetServiceGroups/max_1000_with_100_groups-4             1.21k ± 0%
GetServiceGroups/max_1000_with_200_groups-4             1.21k ± 0%
GetServiceGroups/max_1000_with_300_groups-4             1.21k ± 0%
GetServiceGroups/max_1000_with_400_groups-4             1.21k ± 0%
GetServiceGroups/max_1000_with_500_groups-4             1.21k ± 0%
GetServiceGroups/max_1000_with_600_groups-4             1.21k ± 0%
GetServiceGroups/max_1000_with_700_groups-4             1.21k ± 0%
GetServiceGroups/max_1000_with_800_groups-4             1.21k ± 0%
GetServiceGroups/max_1000_with_900_groups-4             1.21k ± 0%
GetServiceGroups/max_1000_with_1000_groups-4            1.21k ± 0%
GetServiceGroupsHeavyLoad/max_5000_with_0_groups-4        427 ± 0%
GetServiceGroupsHeavyLoad/max_5000_with_1000_groups-4   1.21k ± 0%
GetServiceGroupsHeavyLoad/max_5000_with_2000_groups-4   1.21k ± 0%
GetServiceGroupsHeavyLoad/max_5000_with_3000_groups-4   1.22k ± 0%
GetServiceGroupsHeavyLoad/max_5000_with_4000_groups-4   1.21k ± 0%
GetServiceGroupsHeavyLoad/max_5000_with_5000_groups-4   1.21k ± 0%
GetServices/max_1000_with_0_services-4                    652 ± 0%
GetServices/max_1000_with_100_services-4                1.48k ± 0%
GetServices/max_1000_with_200_services-4                1.48k ± 0%
GetServices/max_1000_with_300_services-4                1.48k ± 0%
GetServices/max_1000_with_400_services-4                1.48k ± 0%
GetServices/max_1000_with_500_services-4                1.48k ± 0%
GetServices/max_1000_with_600_services-4                1.48k ± 0%
GetServices/max_1000_with_700_services-4                1.48k ± 0%
GetServices/max_1000_with_800_services-4                1.48k ± 0%
GetServices/max_1000_with_900_services-4                1.48k ± 0%
GetServices/max_1000_with_1000_services-4               1.48k ± 0%
GetServicesHeavyLoad/max_5000_with_0_services-4           652 ± 0%
GetServicesHeavyLoad/max_5000_with_1000_services-4      1.48k ± 0%
GetServicesHeavyLoad/max_5000_with_2000_services-4      1.48k ± 0%
GetServicesHeavyLoad/max_5000_with_3000_services-4      1.48k ± 0%
GetServicesHeavyLoad/max_5000_with_4000_services-4      1.48k ± 0%
GetServicesHeavyLoad/max_5000_with_5000_services-4      1.48k ± 0%
```